+++ src/Form/BriefingType.php
@@ -2,7 +2,7 @@
 
 namespace App\Form;
 
-use App\Entity\InspectionLetter;
+use App\Entity\HearingBriefing;
 use App\Entity\Party;
 use App\Traits\TemplateFormTrait;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
@@ -14,7 +14,7 @@ use Symfony\Component\Translation\TranslatableMessage;
 use Symfony\Component\Validator\Constraints\Count;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
-class InspectionLetterType extends AbstractType
+class BriefingType extends AbstractType
 {
     use TemplateFormTrait;
 
@@ -25,37 +25,36 @@ class InspectionLetterType extends AbstractType
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'data_class' => InspectionLetter::class,
+            'data_class' => HearingBriefing::class,
+            'case_parties' => null,
             'mail_template_choices' => null,
-            'available_recipients' => null,
+            'preselects' => null,
         ]);
     }
 
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $availableRecipients = $options['available_recipients'];
-        $availableTemplateChoices = $options['mail_template_choices'];
-
         $builder
             ->add('title', TextType::class, [
-                'label' => $this->translator->trans('Title', [], 'agenda'),
-                'help' => $this->translator->trans('Choose a title for the inspection letter', [], 'agenda'),
+                'label' => $this->translator->trans('Title', [], 'case'),
             ])
             ->add('recipients', EntityType::class, [
                 'class' => Party::class,
                 'label' => $this->translator->trans('Recipients', [], 'case'),
-                'choices' => $availableRecipients,
+                'choices' => $options['case_parties'],
                 'choice_label' => function ($key, $value) {
                     return $value;
                 },
                 'multiple' => true,
                 'expanded' => true,
+                'data' => $options['preselects'],
+                'mapped' => false,
                 'constraints' => [
-                    new Count(['min' => 1, 'minMessage' => new TranslatableMessage('Your inspection letter must have at least one recipient', [], 'validators')]),
+                    new Count(['min' => 1, 'minMessage' => new TranslatableMessage('Your briefing letter must have at least one recipient', [], 'validators')]),
                 ],
             ])
         ;
 
-        $this->addTemplate($builder, $availableTemplateChoices);
+        $this->addTemplate($builder, $options['mail_template_choices']);
     }
 }